rename Typography componenty
